requirements.txt
Django==1.8.2
Pillow==2.6.1
Unipath==1.1
arrow==0.5.0
bibtexparser==0.6.0
dj-database-url==0.3.0
django-extensions==1.5.5
dropbox==2.2.0
future==0.14.3
gunicorn==19.3.0
memoized-property==1.0.2
mendeley==0.3.2
oauthlib==0.7.2
pdfminer==20140328
psycopg2==2.6
python-dateutil==2.4.2
python-decouple==2.
python-docx==0.8.5
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.4.2
six==1.9.0
urllib3==1.10.4
wsgiref==0.1.2
beautifulsoup4==4.3.2
